requirements.txt
python-dotenv
langchain
langchain_core
langchain_community
langchain_groq
langchain_google_genai
streamlit
faiss-cpu
langchain_huggingface
PyPDF2
pdfplumber